requirements.txt
Adafruit-Blinka==0.3.2
adafruit-circuitpython-busdevice==2.2.7
adafruit-circuitpython-mcp230xx==1.1.1
Adafruit-GPIO==1.0.3
Adafruit-MCP3008==1.0.2
Adafruit-MotorHAT==1.4.0
Adafruit-PureIO==0.2.3
alabaster==0.7.12
amqp==1.4.9
anyjson==0.3.3
astroid==2.1.0
async-generator==1.10
attrs==18.2.0
Babel==2.6.0
billiard==3.3.0.23
bleach==3.0.2
blessings==1.7
build-utils==0.3.2
celery==3.1.26.post2
certifi==2018.10.15
chardet==3.0.4
colorama==0.4.1
CommonMark==0.5.4
contextvars==2.3
coverage==4.5.2
cssselect==1.0.3
Django==2.1.2
django-celery==3.2.2
django-session-csrf==0.7.1
docopt==0.6.2
docutils==0.14
fake-rpi==0.6.0
flower==0.9.2
gunicorn==19.9.0
idna==2.7
imagesize==1.1.0
immutables==0.6
isort==4.3.4
Jinja2==2.10
kombu==3.0.37
lazy-object-proxy==1.3.1
lxml==4.2.5
MarkupSafe==1.1.0
mccabe==0.6.1
nose==1.3.7
numpy==1.15.4
outcome==1.0.0
packaging==18.0
parsel==1.5.1
Pillow==5.3.0
pkginfo==1.4.2
psycopg2-binary==2.7.5
purepng==0.2.0
Pygments==2.3.0
pylint==2.2.2
pylint-django==2.0.4
pylint-plugin-utils==0.4
pyparsing==2.3.0
pytz==2018.5
readme-renderer==24.0
recommonmark==0.4.0
requestium==0.1.9
requests==2.20.1
requests-file==1.4.3
requests-toolbelt==0.8.0
rinoh-typeface-dejavuserif==0.1.1
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.3.1
rpi-ws281x==4.1.0
RPi.GPIO==0.6.5
selenium==3.141.0
six==1.11.0
sniffio==1.0.0
snowballstemmer==1.2.1
sortedcontainers==2.1.0
soup==0.1.0
sphinxcontrib-websupport==1.1.0
spidev==3.2
tldextract==2.2.0
tornado==5.1.1
tqdm==4.28.1
twine==1.12.1
typed-ast==1.1.0
urllib3==1.24.1
w3lib==1.19.0
webencodings==0.5.1
wget==3.2
wrapt==1.10.11
